Move mock to stubs
